Move default parameter estimation code into FqTemplateBrowser
