+++ qt/scientific_interfaces/Inelastic/Analysis/FunctionBrowser/FqTemplateBrowser.h
@@ -1,0 +1,25 @@
+// Mantid Repository : https://github.com/mantidproject/mantid
+//
+// Copyright &copy; 2023 ISIS Rutherford Appleton Laboratory UKRI,
+//   NScD Oak Ridge National Laboratory, European Spallation Source,
+//   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
+// SPDX - License - Identifier: GPL - 3.0 +
+#pragma once
+
+#include "Analysis/FunctionBrowser/SingleFunctionTemplateBrowser.h"
+#include "DllConfig.h"
+
+namespace MantidQt {
+namespace CustomInterfaces {
+namespace IDA {
+
+class MANTIDQT_INELASTIC_DLL FqTemplateBrowser : public SingleFunctionTemplateBrowser {
+  Q_OBJECT
+
+public:
+  FqTemplateBrowser();
+};
+
+} // namespace IDA
+} // namespace CustomInterfaces
+} // namespace MantidQt
